[H1] kwon: 헬스리스트 체크박스 코드 정리, 컴포넌트별 상태 분배
