+++ rootfit_front/src/todo/sections/view/TodoMain.jsx
@@ -10,7 +10,6 @@ import ThemeProvider from '../../theme';
 
 // context api
 import UserContext from '../../../user/context/UserContext';
-import TodoContext from '../../context/todoContext';
 
 //-------------------------------------------------------------------
 
@@ -19,10 +18,6 @@ const TodoContainer = () => {
   const values = useContext(UserContext);
   const userID = values.state.user.id;
 
-  // 헬스리스트 Context
-  const todoValues = useContext(TodoContext);
-  const todoActions = todoValues.actions;
-
   const navigate = useNavigate();
 
   // 회원 여부 체크
@@ -30,12 +25,10 @@ const TodoContainer = () => {
     const currentUrl = '/todo';
     //로그인하지 않은 경우
     if (userID === '') {
-      //로그인 페이지에서 로그인후 다시 보고있던 페이지로 원복 -> SignIn.jsx에 redirect 연계해야함.
       navigate(`/user/signin?redirect=${encodeURIComponent(currentUrl)}`);
-      return;
     } else {
       //로그인한 경우
-      todoActions.getLoadSelect(userID);
+      navigate(currentUrl);
     }
   });
